Clock Widget Tutorial
This is Part I of a project to build a timer widget
It follows a tutorial (url below) to create a clock
widget. My goal is to familiarize myself with some
of the differences between developing widgets, given
my experience developing apps.

From this base, I am hoping to expand the code to
perform countdowns and countups to various datetimes
set by the user.

https://code.tutsplus.com/tutorials/building-a-customizable-android-analog-clock-widget--mobile-11685
Thanks to Sue Smith for a great introduction to Widget development
I'm really excited for how simple it was to bridge into Widgets!