Clock Widget Tutorial - Addendum
This expands on Part I of a project to build a timer widget
It adds a digital clock to the previous 3 analog clocks the
tutorial set up. My goal is to test my understanding enough
to add a digital readout, which is more in line with timers
that the project eventually needs.

The URL of the original tutorial is again:
https://code.tutsplus.com/tutorials/building-a-customizable-android-analog-clock-widget--mobile-11685